Add deploy script for making npm deploys easier
This adds a bash script which looks for common mistakes before running deploys. It looks to ensure all versions match, that git has pulled, that it's not on a feature branch, and a tag has been applied to the current comment. It also determines if it is a pre-release publish.